update: remove certain type converting warnings
git-svn-id: https://www.h3d.org:8090/H3DAPI/trunk/H3DAPI@3128 610dfd11-2700-0410-9b68-9ae28eac5983

+++ src/X3DTexture2DNode.cpp
@@ -170,12 +170,12 @@ void X3DTexture2DNode::glTexImage( Image *i, GLenum _texture_target,
     unsigned int new_height = i->height(); 
 
     // Enforce global maximum texture dimension
-    if ( max_dimension > 0 && new_width > max_dimension ) {
-      new_width= max_dimension;
+    if ( max_dimension > 0 && new_width > (unsigned int)max_dimension ) {
+      new_width= (unsigned int)max_dimension;
       needs_scaling= true;
     }
-    if ( max_dimension > 0 && new_height > max_dimension ) {
-      new_height= max_dimension;
+    if ( max_dimension > 0 && new_height > (unsigned int)max_dimension ) {
+      new_height= (unsigned int)max_dimension;
       needs_scaling= true;
     }
